+++ server/src/main/java/Server/ServerCommunicator.java
@@ -53,7 +53,7 @@ public class ServerCommunicator {
             port = SERVER_PORT_NUMBER;
         }
 
-        System.out.println("Starting Full Server on port " + port + "...");
+        System.out.println("Starting Ticket To Ride Server on port " + port + "...");
 
         try {
             server = HttpServer.create(new InetSocketAddress(port), MAX_WAITING_CONNECTIONS);
@@ -99,43 +99,6 @@ public class ServerCommunicator {
         // 1. Return the max number of server connections
 
         return MAX_WAITING_CONNECTIONS;
-    }
-
-    /**
-     * Attempts to turn the String into an integer. Returns a 0 if this is not possible.
-     *
-     * @param intString the String to convert to an integer.
-     *
-     * @return the parameter converted to an integer, or a zero if this is not possible.
-     */
-    public static int parseInt(String intString){
-        int returnValue = 0;
-
-        try {
-            returnValue = Integer.parseInt(intString);
-        }
-        catch (Exception exc){
-            returnValue = 0;
-        }
-
-        return returnValue;
-    }
-
-    /**
-     * Starts the server
-     *
-     * @param args the arguements for the server
-     */
-    public static void main(String[] args){
-        // Steps in method:
-        // 1. Start the server
-
-        if (args.length > 0){
-            new ServerCommunicator().run(parseInt(args[0]));
-        }
-        else {
-            new ServerCommunicator().run();
-        }
     }
 
 }
